test: move all dbus custom interfaces in `ubo_app.utils.dbus_interfaces` and preserve it in test cleanup, this is due to sdbus having a hidden mapping of dbus interfaces to their implementations in the C code and it doesn't clean up the mapping when the interface is removed
